requirements.txt
flask==3.1.0
gunicorn==23.0.0
anthropic==0.42.0
weasyprint==63.1
python-docx==1.1.2
openpyxl==3.1.5
PyPDF2==3.0.1
python-dotenv==1.0.1
requests==2.32.3
